move : 쿠폰 도메인 이동시 생긴 테스트 코드 수정
