+++ src/main/java/bc1/gream/domain/buy/provider/BuyBidProvider.java
@@ -8,10 +8,10 @@ import bc1.gream.domain.buy.mapper.BuyMapper;
 import bc1.gream.domain.buy.repository.BuyRepository;
 import bc1.gream.domain.buy.service.BuyService;
 import bc1.gream.domain.common.facade.ChangingCouponStatusFacade;
+import bc1.gream.domain.coupon.entity.CouponStatus;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.sell.service.helper.deadline.Deadline;
 import bc1.gream.domain.sell.service.helper.deadline.DeadlineCalculator;
-import bc1.gream.domain.user.coupon.entity.CouponStatus;
 import bc1.gream.domain.user.entity.User;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
